requirements.txt
fastapi>=0.115
uvicorn>=0.34
pywebview>=5.0
rumps>=0.4
